--- android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/onboarding/ActivityCategory.kt
@@ -1,7 +1,0 @@
-package com.emmsale.presentation.ui.onboarding
-
-enum class ActivityCategory(val title: String) {
-    EDUCATION("교육"),
-    CLUB("동아리"),
-    FIELD("직무"),
-}
