+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/onboarding/education/OnboardingEducationFragment.kt
@@ -1,40 +1,42 @@
-package com.emmsale.presentation.ui.onboarding
+package com.emmsale.presentation.ui.onboarding.education
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
 import com.emmsale.R
-import com.emmsale.databinding.FragmentOnboardingFieldBinding
+import com.emmsale.databinding.FragmentOnboardingEducationBinding
 import com.emmsale.presentation.base.fragment.BaseFragment
 import com.emmsale.presentation.common.views.activityChipOf
+import com.emmsale.presentation.ui.onboarding.OnboardingActivity
+import com.emmsale.presentation.ui.onboarding.OnboardingViewModel
 import com.emmsale.presentation.ui.onboarding.uistate.ActivityUiState
 
-class OnboardingFieldFragment : BaseFragment<FragmentOnboardingFieldBinding>() {
+class OnboardingEducationFragment :
+    BaseFragment<FragmentOnboardingEducationBinding>(),
+    View.OnClickListener {
     val viewModel: OnboardingViewModel by activityViewModels { OnboardingViewModel.factory }
-    override val layoutResId: Int = R.layout.fragment_onboarding_field
+    override val layoutResId: Int = R.layout.fragment_onboarding_education
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.viewModel = viewModel
         initClickListener()
-        setupFields()
+        setupEducations()
     }
 
     private fun initClickListener() {
-        binding.btnNext.setOnClickListener {
-            (requireActivity() as OnboardingActivity).navigateToNextPage()
-        }
+        binding.btnNext.setOnClickListener(this)
     }
 
-    private fun setupFields() {
+    private fun setupEducations() {
         viewModel.activities.observe(viewLifecycleOwner) { activities ->
-            binding.chipgroupFieldTags.removeAllViews()
-            activities.fields.forEach(::addFieldChip)
+            binding.chipgroupEduTags.removeAllViews()
+            activities.educations.forEach(::addEducationChip)
         }
     }
 
-    private fun addFieldChip(fieldTag: ActivityUiState) {
-        binding.chipgroupFieldTags.addView(createChip(fieldTag))
+    private fun addEducationChip(educationTag: ActivityUiState) {
+        binding.chipgroupEduTags.addView(createChip(educationTag))
     }
 
     private fun createChip(activity: ActivityUiState) = activityChipOf {
@@ -42,6 +44,12 @@ class OnboardingFieldFragment : BaseFragment<FragmentOnboardingFieldBinding>() {
         isChecked = activity.isSelected
         setOnCheckedChangeListener { _, isChecked ->
             viewModel.updateSelection(activity.id, isChecked)
+        }
+    }
+
+    override fun onClick(view: View) {
+        when (view.id) {
+            R.id.btn_next -> (requireActivity() as OnboardingActivity).navigateToNextPage()
         }
     }
 }
