+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/onboarding/field/OnboardingFieldFragment.kt
@@ -1,40 +1,42 @@
-package com.emmsale.presentation.ui.onboarding
+package com.emmsale.presentation.ui.onboarding.field
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
 import com.emmsale.R
-import com.emmsale.databinding.FragmentOnboardingEducationBinding
+import com.emmsale.databinding.FragmentOnboardingFieldBinding
 import com.emmsale.presentation.base.fragment.BaseFragment
 import com.emmsale.presentation.common.views.activityChipOf
+import com.emmsale.presentation.ui.onboarding.OnboardingActivity
+import com.emmsale.presentation.ui.onboarding.OnboardingViewModel
 import com.emmsale.presentation.ui.onboarding.uistate.ActivityUiState
 
-class OnboardingEducationFragment :
-    BaseFragment<FragmentOnboardingEducationBinding>(),
-    View.OnClickListener {
+class OnboardingFieldFragment : BaseFragment<FragmentOnboardingFieldBinding>() {
     val viewModel: OnboardingViewModel by activityViewModels { OnboardingViewModel.factory }
-    override val layoutResId: Int = R.layout.fragment_onboarding_education
+    override val layoutResId: Int = R.layout.fragment_onboarding_field
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.viewModel = viewModel
         initClickListener()
-        setupEducations()
+        setupFields()
     }
 
     private fun initClickListener() {
-        binding.btnNext.setOnClickListener(this)
-    }
-
-    private fun setupEducations() {
-        viewModel.activities.observe(viewLifecycleOwner) { activities ->
-            binding.chipgroupEduTags.removeAllViews()
-            activities.educations.forEach(::addEducationChip)
+        binding.btnNext.setOnClickListener {
+            (requireActivity() as OnboardingActivity).navigateToNextPage()
         }
     }
 
-    private fun addEducationChip(educationTag: ActivityUiState) {
-        binding.chipgroupEduTags.addView(createChip(educationTag))
+    private fun setupFields() {
+        viewModel.activities.observe(viewLifecycleOwner) { activities ->
+            binding.chipgroupFieldTags.removeAllViews()
+            activities.fields.forEach(::addFieldChip)
+        }
+    }
+
+    private fun addFieldChip(fieldTag: ActivityUiState) {
+        binding.chipgroupFieldTags.addView(createChip(fieldTag))
     }
 
     private fun createChip(activity: ActivityUiState) = activityChipOf {
@@ -42,12 +44,6 @@ class OnboardingEducationFragment :
         isChecked = activity.isSelected
         setOnCheckedChangeListener { _, isChecked ->
             viewModel.updateSelection(activity.id, isChecked)
-        }
-    }
-
-    override fun onClick(view: View) {
-        when (view.id) {
-            R.id.btn_next -> (requireActivity() as OnboardingActivity).navigateToNextPage()
         }
     }
 }
